Add release creation for when release doesn't exist
If you create a tag and push it to GitHub, it doesn't automatically
make a release. The tag shows up under the releases menu with the
.tag.gz and .zip archives of the source, but you can't add assets
through the API without making it a "real" release. So if a release for
the tag can't be found via the API, we try to create the release. If
the tag already exists, then the GitHub API makes the release from
'tag_name', which is what we want.